rename: Contact 폴더 이름 수정 #65
