+++ src/routes/PublicRoutes.tsx
@@ -11,7 +11,7 @@ import PostDetail from '@/pages/Community/PostDetail';
 import Project from '@/pages/Community/Project';
 import Qna from '@/pages/Community/Qna';
 import Submit from '@/pages/Community/Submit';
-import Contact from '@/pages/contact';
+import Contact from '@/pages/Contact';
 import Home from '@/pages/Home';
 
 export const publicRoutes: RouteObject[] = [
